uniform exceptions (#29)

+++ src/src/BasaltHexagons.UniversalFileSystem.File/FileFileSystem.cs
@@ -6,6 +6,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using BasaltHexagons.UniversalFileSystem.Core;
 using BasaltHexagons.UniversalFileSystem.Core.Disposing;
+using BasaltHexagons.UniversalFileSystem.Core.Exceptions;
 
 namespace BasaltHexagons.UniversalFileSystem.File;
 
@@ -14,7 +15,8 @@ class FileFileSystem : AsyncDisposable, IFileSystem
 {
     #region IFileSystem
 
-    public async IAsyncEnumerable<ObjectMetadata> ListObjectsAsync(Uri prefix, bool recursive, [EnumeratorCancellation] CancellationToken cancellationToken)
+    public async IAsyncEnumerable<ObjectMetadata> ListObjectsAsync(Uri prefix, bool recursive,
+        [EnumeratorCancellation] CancellationToken cancellationToken)
     {
         EnumerationOptions enumerationOptions = new()
         {
@@ -28,7 +30,8 @@ class FileFileSystem : AsyncDisposable, IFileSystem
             {
                 cancellationToken.ThrowIfCancellationRequested();
 
-                ObjectMetadata? metadata = await this.GetObjectMetadataInternalAsync(new Uri(entry), false, cancellationToken);
+                ObjectMetadata? metadata =
+                    await this.GetObjectMetadataInternalAsync(new Uri(entry), false, cancellationToken);
                 if (metadata != null)
                     yield return metadata;
             }
@@ -37,7 +40,8 @@ class FileFileSystem : AsyncDisposable, IFileSystem
             {
                 cancellationToken.ThrowIfCancellationRequested();
 
-                ObjectMetadata? metadata = await this.GetObjectMetadataInternalAsync(new Uri(entry), true, cancellationToken);
+                ObjectMetadata? metadata =
+                    await this.GetObjectMetadataInternalAsync(new Uri(entry), true, cancellationToken);
                 if (metadata != null)
                     yield return metadata;
 
@@ -70,22 +74,28 @@ class FileFileSystem : AsyncDisposable, IFileSystem
         }
     }
 
-    public Task<ObjectMetadata?> GetFileMetadataAsync(Uri uri, CancellationToken cancellationToken)
+    public Task<ObjectMetadata> GetFileMetadataAsync(Uri uri, CancellationToken cancellationToken)
     {
         return this.GetObjectMetadataInternalAsync(uri, false, cancellationToken);
     }
 
-    public Task<Stream> GetFileAsync(Uri uri, CancellationToken cancellationToken)
+    public async Task<Stream> GetFileAsync(Uri uri, CancellationToken cancellationToken)
     {
-        return Task.FromResult((Stream)new FileStream(uri.AbsolutePath, FileMode.Open, FileAccess.Read));
+        if (!await this.DoesFileExistAsync(uri, cancellationToken))
+            throw new FileNotExistsException(uri);
+        return new FileStream(uri.AbsolutePath, FileMode.Open, FileAccess.Read);
     }
 
     public async Task PutFileAsync(Uri uri, Stream stream, bool overwrite, CancellationToken cancellationToken)
     {
+        if (!overwrite && await this.DoesFileExistAsync(uri, cancellationToken))
+            throw new FileExistsException(uri);
+
         string? dir = Path.GetDirectoryName(uri.AbsolutePath);
         if (dir != null)
             Directory.CreateDirectory(dir);
-        await using FileStream fileStream = new(uri.AbsolutePath, overwrite ? FileMode.Create : FileMode.CreateNew, FileAccess.Write);
+        await using FileStream fileStream = new(uri.AbsolutePath, overwrite ? FileMode.Create : FileMode.CreateNew,
+            FileAccess.Write);
         await stream.CopyToAsync(fileStream, cancellationToken);
     }
 
@@ -98,47 +108,61 @@ class FileFileSystem : AsyncDisposable, IFileSystem
         return true;
     }
 
-    public Task MoveFileAsync(Uri oldPath, Uri newPath, bool overwrite, CancellationToken cancellationToken)
+    public async Task MoveFileAsync(Uri oldPath, Uri newPath, bool overwrite, CancellationToken cancellationToken)
     {
         if (oldPath == newPath)
             throw new ArgumentException("Can't move file to itself.");
 
+        if (!await this.DoesFileExistAsync(oldPath, cancellationToken))
+            throw new FileNotExistsException(oldPath);
+
+        if (!overwrite && await this.DoesFileExistAsync(newPath, cancellationToken))
+            throw new FileExistsException(newPath);
+
         string? directory = Path.GetDirectoryName(newPath.AbsolutePath);
         if (directory == null)
-            return Task.FromException(new ArgumentException($"Can't get directory from uri {newPath}"));
+            throw new ArgumentException($"Can't get directory from uri {newPath}");
         Directory.CreateDirectory(directory);
 
         System.IO.File.Move(oldPath.AbsolutePath, newPath.AbsolutePath, overwrite);
-        return Task.CompletedTask;
     }
 
-    public Task CopyFileAsync(Uri sourceUri, Uri destUri, bool overwrite, CancellationToken cancellationToken)
+    public async Task CopyFileAsync(Uri sourceUri, Uri destUri, bool overwrite, CancellationToken cancellationToken)
     {
         if (sourceUri == destUri)
             throw new ArgumentException("Can't copy file to itself.");
 
+        if (!await this.DoesFileExistAsync(sourceUri, cancellationToken))
+            throw new FileNotExistsException(sourceUri);
+
+        if (!overwrite && await this.DoesFileExistAsync(destUri, cancellationToken))
+            throw new FileExistsException(destUri);
+
         string? directory = Path.GetDirectoryName(destUri.AbsolutePath);
         if (directory == null)
-            return Task.FromException(new ArgumentException($"Can't get directory from uri {destUri}"));
+            throw new ArgumentException($"Can't get directory from uri {destUri}");
         Directory.CreateDirectory(directory);
 
         System.IO.File.Copy(sourceUri.AbsolutePath, destUri.AbsolutePath, overwrite);
-        return Task.CompletedTask;
     }
 
-    public Task<bool> DoesFileExistAsync(Uri uri, CancellationToken cancellationToken) => Task.FromResult(System.IO.File.Exists(uri.AbsolutePath));
+    public Task<bool> DoesFileExistAsync(Uri uri, CancellationToken cancellationToken) =>
+        Task.FromResult(System.IO.File.Exists(uri.AbsolutePath));
 
     #endregion
 
-    private Task<bool> DoesDirectoryExistAsync(Uri uri, CancellationToken cancellationToken) => Task.FromResult(System.IO.Directory.Exists(uri.AbsolutePath));
+    private Task<bool> DoesDirectoryExistAsync(Uri uri, CancellationToken cancellationToken) =>
+        Task.FromResult(System.IO.Directory.Exists(uri.AbsolutePath));
 
-    private async Task<ObjectMetadata?> GetObjectMetadataInternalAsync(Uri uri, bool returnDirectory, CancellationToken cancellationToken)
+    private async Task<ObjectMetadata> GetObjectMetadataInternalAsync(Uri uri, bool returnDirectory,
+        CancellationToken cancellationToken)
     {
         if (await this.DoesFileExistAsync(uri, cancellationToken))
-            return new ObjectMetadata(uri, ObjectType.File, new FileInfo(uri.AbsolutePath).Length, System.IO.File.GetLastWriteTimeUtc(uri.AbsolutePath));
+            return new ObjectMetadata(uri, ObjectType.File, new FileInfo(uri.AbsolutePath).Length,
+                System.IO.File.GetLastWriteTimeUtc(uri.AbsolutePath));
         if (returnDirectory && await this.DoesDirectoryExistAsync(uri, cancellationToken))
             return new ObjectMetadata(new Uri($"{uri}/"), ObjectType.Prefix, null, null);
 
-        return null;
+        throw new FileNotExistsException(uri);
     }
 }
